+++ Trackmeal/Data/Migrations/ApplicationDbContextModelSnapshot.cs
@@ -235,8 +235,8 @@ namespace Trackmeal.Data.Migrations
                     b.Property<byte>("Amount")
                         .HasColumnType("tinyint");
 
-                    b.Property<long?>("OrderId")
-                        .HasColumnType("bigint");
+                    b.Property<int?>("OrderId")
+                        .HasColumnType("int");
 
                     b.Property<int>("ProductId")
                         .HasColumnType("int");
@@ -252,11 +252,11 @@ namespace Trackmeal.Data.Migrations
 
             modelBuilder.Entity("Trackmeal.Models.Order", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint");
+                        .HasColumnType("int");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"), 1L, 1);
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"), 1L, 1);
 
                     b.Property<DateTime>("DateOrdered")
                         .HasColumnType("datetime2");
